chore: add compiled executable in ZIP
This commit adds a ZIP file containing the compiled version of the game using PyInstaller, allowing users to easily download and run the game without needing to set up a Python environment. This makes the game more accessible to others who want to try it without the need to install dependencies or run the source code.